leixperiment sync

+++ code/ui/ui_syscalls.c
@@ -30,9 +30,15 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 static intptr_t (QDECL *syscall)( intptr_t arg, ... ) = (intptr_t (QDECL *)( intptr_t, ...))-1;
 
-void dllEntry( intptr_t (QDECL *syscallptr)( intptr_t arg,... ) ) {
+void dllEntry( long (QDECL *syscallptr)( long arg,... ) ) {
 	syscall = syscallptr;
 }
+
+//static intptr_t (QDECL *syscall)( intptr_t arg, ... ) = (intptr_t (QDECL *)( intptr_t, ...))-1;
+
+//void dllEntry( intptr_t (QDECL *syscallptr)( intptr_t arg,... ) ) {
+	//syscall = syscallptr;
+//}
 
 int PASSFLOAT( float x ) {
 	float	floatTemp;
